% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.r
\name{as.sr}
\alias{as.sr}
\alias{as.sr.default}
\alias{as.sr.matrix}
\alias{as.sr.data.frame}
\alias{as.sr.lm}
\alias{as.sr.xts}
\alias{as.sr.timeSeries}
\title{Compute the Sharpe ratio.}
\usage{
as.sr(x, c0 = 0, ope = 1, na.rm = FALSE, epoch = "yr", higher_order = FALSE)

\method{as.sr}{default}(x, c0 = 0, ope = 1, na.rm = FALSE, epoch = "yr", higher_order = FALSE)

\method{as.sr}{matrix}(x, c0 = 0, ope = 1, na.rm = FALSE, epoch = "yr", higher_order = FALSE)

\method{as.sr}{data.frame}(x, c0 = 0, ope = 1, na.rm = FALSE, epoch = "yr", higher_order = FALSE)

\method{as.sr}{lm}(x, c0 = 0, ope = 1, na.rm = FALSE, epoch = "yr", higher_order = FALSE)

\method{as.sr}{xts}(x, c0 = 0, ope = 1, na.rm = FALSE, epoch = "yr", higher_order = FALSE)

\method{as.sr}{timeSeries}(x, c0 = 0, ope = 1, na.rm = FALSE, epoch = "yr", higher_order = FALSE)
}
\arguments{
\item{x}{vector of returns, or object of class \code{data.frame}, \code{xts},
or \code{lm}.}

\item{c0}{the 'risk-free' or 'disastrous' rate of return. this is
assumed to be given in the same units as x, \emph{not}
in 'annualized' terms.}

\item{ope}{the number of observations per 'epoch'. For convenience of
interpretation, The Sharpe ratio is typically quoted in 'annualized' 
units for some epoch, that is, 'per square root epoch', though returns 
are observed at a frequency of \code{ope} per epoch. 
The default value is 1, meaning the code will not attempt to guess
what the observation frequency is, and no annualization adjustments
will be made.}

\item{na.rm}{logical.  Should missing values be removed?}

\item{epoch}{the string representation of the 'epoch', defaulting
to 'yr'.}

\item{higher_order}{a Boolean. If true, we compute 
cumulants of the returns to leverage higher order accuracy formulae
when possible.}
}
\value{
a list containing the following components:
\describe{
\item{sr}{the annualized Sharpe ratio.}
\item{df}{the t-stat degrees of freedom.}
\item{c0}{the risk free term.}
\item{ope}{the annualization factor.}
\item{rescal}{the rescaling factor.}
\item{epoch}{the string epoch.}
}
cast to class \code{sr}.
}
\description{
Computes the Sharpe ratio of some observed returns.
}
\details{
Suppose \eqn{x_i}{xi} are \eqn{n}{n} independent returns of some
asset.
Let \eqn{\bar{x}}{xbar} be the sample mean, and \eqn{s}{s} be
the sample standard deviation (using Bessel's correction). Let \eqn{c_0}{c0}
be the 'risk free rate'.  Then
\deqn{z = \frac{\bar{x} - c_0}{s}}{z = (xbar - c0)/s} 
is the (sample) Sharpe ratio.

The units of \eqn{z}{z} are \eqn{\mbox{time}^{-1/2}}{per root time}.
Typically the Sharpe ratio is \emph{annualized} by multiplying by
\eqn{\sqrt{\mbox{ope}}}{sqrt(ope)}, where \eqn{\mbox{ope}}{ope} 
is the number of observations
per year (or whatever the target annualization epoch.)

Note that if \code{ope} is not given, the converter from \code{xts}
attempts to infer the observations per year, without regard to 
the name of the \code{epoch} given.
}
\examples{
# Sharpe's 'model': just given a bunch of returns.
asr <- as.sr(rnorm(253*3),ope=253)
# or a matrix, with a name
my.returns <- matrix(rnorm(253*3),ncol=1)
colnames(my.returns) <- c("my strategy")
asr <- as.sr(my.returns)

# given an xts object:
if (require(xts)) {
 data(stock_returns)
 IBM <- stock_returns[,'IBM']
 asr <- as.sr(IBM,na.rm=TRUE)
}

# on a linear model, find the 'Sharpe' of the residual term
nfac <- 5
nyr <- 10
ope <- 253
set.seed(as.integer(charToRaw("determinstic")))
Factors <- matrix(rnorm(ope*nyr*nfac,mean=0,sd=0.0125),ncol=nfac)
Betas <- exp(0.1 * rnorm(dim(Factors)[2]))
Returns <- (Factors \%*\% Betas) + rnorm(dim(Factors)[1],mean=0.0005,sd=0.012)
APT_mod <- lm(Returns ~ Factors)
asr <- as.sr(APT_mod,ope=ope)
# try again, but make the Returns independent of the Factors.
Returns <- rnorm(dim(Factors)[1],mean=0.0005,sd=0.012)
APT_mod <- lm(Returns ~ Factors)
asr <- as.sr(APT_mod,ope=ope)

# compute the Sharpe of a bunch of strategies:
my.returns <- matrix(rnorm(253*3*4),ncol=4)
asr <- as.sr(my.returns)  # without sensible colnames?
colnames(my.returns) <- c("strat a","strat b","strat c","strat d")
asr <- as.sr(my.returns)
  
}
\references{
Sharpe, William F. "Mutual fund performance." Journal of business (1966): 119-138.
\url{https://ideas.repec.org/a/ucp/jnlbus/v39y1965p119.html}

Lo, Andrew W. "The statistics of Sharpe ratios." Financial Analysts Journal 58, no. 4 
(2002): 36-52. \url{https://www.ssrn.com/paper=377260}
}
\seealso{
\code{\link{reannualize}}

sr-distribution functions, \code{\link{dsr}, \link{psr}, \link{qsr}, \link{rsr}}

Other sr: 
\code{\link{confint.sr}()},
\code{\link{dsr}()},
\code{\link{is.sr}()},
\code{\link{plambdap}()},
\code{\link{power.sr_test}()},
\code{\link{predint}()},
\code{\link{print.sr}()},
\code{\link{reannualize}()},
\code{\link{se}()},
\code{\link{sr}},
\code{\link{sr_equality_test}()},
\code{\link{sr_test}()},
\code{\link{sr_unpaired_test}()},
\code{\link{sr_vcov}()},
\code{\link{summary.sr}}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\concept{sr}
\keyword{univar}
