% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netLoadAssembly.R
\name{netLoadAssembly}
\alias{netLoadAssembly}
\title{Load assembly.}
\usage{
netLoadAssembly(filePath)
}
\arguments{
\item{filePath}{Assembly file. It can be the full file path of the assembly, or a qualified assembly name.}
}
\description{
Loads an assembly in the Clr (Common Language Runtime).
}
\examples{
\dontrun{
library(sharper)

pkgPath <- path.package("sharper")
f <- file.path(pkgPath, "tests", "AssemblyForTests.dll")
netLoadAssembly(f)
}
}
