% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netNew.R
\name{netNew}
\alias{netNew}
\title{Instanciate a .Net object}
\usage{
netNew(typeName, ...)
}
\arguments{
\item{typeName}{The .Net full name type}

\item{...}{.Net Constructor arguments.}
}
\value{
Returns a converted .Net instance if a converter is defined, an external pointer otherwise.
}
\description{
Instanciate a .Net object from its type name.
}
\details{
The \code{typeName} should respect the full type name convention: \code{Namespace.TypeName}
Ellipses \code{...} has to keep the .net constructor arguments order, the named arguments are not supported yet.
If there is many constructors defined for the given .Net type, a score selection is computed from your arguments orders and types to choose the best one.
We consider as a higher priority single value compare to collection of values.
}
\examples{
\dontrun{
library(sharper)

package_folder <- path.package("sharper")
netLoadAssembly(file.path(package_folder, "tests", "RDotNet.AssemblyTest.dll"))

x <- netNew("RDotNet.AssemblyTest.OneCtorData", 21L)
netCall(x, "ToString")
}
}
