% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netSet.R
\name{netSet}
\alias{netSet}
\title{Set property value}
\usage{
netSet(x, propertyName, value)
}
\arguments{
\item{x}{.Net object}

\item{propertyName}{Property name to set value}

\item{value}{value to set.}
}
\description{
Sets a property value from a .Net object
}
\details{
Allows you to set a property value of a .Net object.
The input value will be converted from R type to a .Net type.

If the property value isn't a native C# type or a mapped conversion type,
you have to use an \code{externalptr} on .Net object or a \verb{NetObject R6} instance.

You can define custom converters in C# for that see \code{RDotNetConverter} C# class.
}
\examples{
\dontrun{
library(sharper)

package_folder = path.package("sharper")

x <- netNew("AssemblyForTests.DefaultCtorData")
netSet(x, "Name", "foo")
x_name <- netGet(x, "Name")

netSet(x, "Integers", c(12L, 23L))
x_integers <- netGet(x, "Integers")

}
}
