% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netObject.R
\name{NetType}
\alias{NetType}
\title{NetType class}
\description{
NetType provide the name and namespace of a .Net type
}
\section{Properties}{

\itemize{
\item \code{Name}: .Net type
\item \code{Namespace}: .Net type namespace
\item \code{FullName}: Full .Net type
}
}

\examples{
\dontrun{
library(sharper)
package_folder <- path.package("sharper")
netLoadAssembly(file.path(package_folder, "tests", "AssemblyForTests.dll"))

netType <- NetType$new("AssemblyForTests", "ManyCtorData")
print(sprintf("Name=\%s, Namespace=\%s, FullName=\%s", 
  netType$Name, netType$Namespace, netType$FullName))

object <- netType$createObject(21L)
object$get("Id")
}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{Name}}{.Net type name.}

\item{\code{Namespace}}{namespace of .Net type.}

\item{\code{FullName}}{.Net type name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NetType$new()}}
\item \href{#method-createObject}{\code{NetType$createObject()}}
\item \href{#method-clone}{\code{NetType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new NetType object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetType$new(namespace, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{namespace}}{Namespace namespace of .Net type.}

\item{\code{name}}{.Net type name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createObject"></a>}}
\subsection{Method \code{createObject()}}{
Create a new .Net object and wrap it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetType$createObject(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ctor arguments of the .Net type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NetType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
