% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netUnwrap.R
\name{netUnwrap}
\alias{netUnwrap}
\title{Unwrap \code{NetObject} to \code{externalptr}}
\usage{
netUnwrap(object)
}
\arguments{
\item{object}{\code{NetObject} instance or a \code{list}.}
}
\value{
Returns the unwrapped \verb{R6 NetObject} instance to .Net \code{externalptr}.
}
\description{
Unwrap a \code{R6} \code{NetObject} instance to a .Net object \code{externalptr}.
}
\details{
todo
}
\examples{
\dontrun{
library(sharper)

package_folder <- path.package("sharper")
netLoadAssembly(file.path(package_folder, "tests", "AssemblyForTests.dll"))

x <- netNew("AssemblyForTests.OneCtorData", 21L)
object <- NetObject$new(ptr = x)
ptr <- netUnwrap(object)

l <- list(
  object, 
  c(12, 13), 
  NetObject$new(typeName = "AssemblyForTests.DefaultCtorData"),
  netNew("AssemblyForTests.DefaultCtorData"))
ptrs <- netUnwrap(l)
}
}
