% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netWrap.R
\name{netWrap}
\alias{netWrap}
\title{Wrap \code{externalptr} into \code{NetObject}}
\usage{
netWrap(ptr)
}
\arguments{
\item{ptr}{a .Net object \code{externalptr} or a \code{list}.}
}
\value{
Returns the wrapped .Net \code{externalptr} into \verb{R6 NetObject} instance.
}
\description{
Wrap .Net object \code{externalptr} into a \code{R6} \code{NetObject} instance.
}
\details{
todo
}
\examples{
\dontrun{
library(sharper)

package_folder <- path.package("sharper")
netLoadAssembly(file.path(package_folder, "tests", "AssemblyForTests.dll"))

x <- netNew("AssemblyForTests.OneCtorData", 21L)
object <- netWrap(x)

l <- list(
  x, 
  c(12, 13), 
  NetObject$new(ptr = netNew("AssemblyForTests.DefaultCtorData")),
  netNew("AssemblyForTests.DefaultCtorData"))
objects <- netWrap(l)
}
}
