\name{print.sr}
\alias{print.sr}
\alias{print.sropt}
\title{Print values.}
\usage{
  \method{print}{sr} (x)

  \method{print}{sropt} (x)
}
\arguments{
  \item{x}{an object of class \code{sr} or \code{sropt}.}
}
\value{
  the object, wrapped in \code{invisible}.
}
\description{
  Displays an object, returning it \emph{invisibly}, (via
  \code{invisible(x)}.)
}

