% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.r
\name{se}
\alias{se}
\alias{se.sr}
\title{Standard error computation}
\usage{
se(z, type)

\method{se}{sr}(z, type = c("t", "Lo", "Mertens", "Bao"))
}
\arguments{
\item{z}{an observed Sharpe ratio statistic, of class \code{sr}.}

\item{type}{estimator type. one of \code{"t", "Lo", "Mertens", "Bao"}}
}
\value{
an estimate of standard error.
}
\description{
Estimates the standard error of the Sharpe ratio statistic.
}
\details{
For an observed Sharpe ratio, estimate the standard error.
The following methods are recognized:

\describe{
\item{t}{The default, based on Johnson & Welch, with a correction
for small sample size. Also known as \code{'Lo'}.}
\item{Mertens}{An approximation to the standard error taking into
skewness and kurtosis of the returns distribution.}
\item{Bao}{An even higher accuracty approximation using higher order
moments.}
}

There should be very little difference between these except for very small
sample sizes.

See \sQuote{The Sharpe Ratio: Statistics and Applications},
sections 2.5.1 and 3.2.3.
}
\note{
The units of the standard error are consistent with those of the
input \code{sr} object.
}
\examples{
asr <- as.sr(rnorm(128,0.2))
anse <- se(asr,type="t")
anse <- se(asr,type="Lo")

}
\references{
Sharpe, William F. "Mutual fund performance." Journal of business (1966): 119-138.
\url{https://ideas.repec.org/a/ucp/jnlbus/v39y1965p119.html}

Johnson, N. L., and Welch, B. L. "Applications of the non-central t-distribution."
Biometrika 31, no. 3-4 (1940): 362-389. \doi{10.1093/biomet/31.3-4.362}

Lo, Andrew W. "The statistics of Sharpe ratios." Financial Analysts Journal 58, no. 4 
(2002): 36-52. \url{https://www.ssrn.com/paper=377260}

Bao, Yong. "Estimation Risk-Adjusted Sharpe Ratio and Fund Performance Ranking 
Under a General Return Distribution." Journal of Financial Econometrics 7, 
no. 2 (2009): 152-173. \doi{10.1093/jjfinec/nbn022}

Opdyke, J. D. "Comparing Sharpe Ratios: So Where are the p-values?" Journal of Asset
Management 8, no. 5 (2006): 308-336. \url{https://www.ssrn.com/paper=886728}

Pav, S. E. "The Sharpe Ratio: Statistics and Applications." CRC Press, 2021.

Walck, C. "Hand-book on STATISTICAL DISTRIBUTIONS for experimentalists."
1996. \url{http://www.stat.rice.edu/~dobelman/textfiles/DistributionsHandbook.pdf}
}
\seealso{
sr-distribution functions, \code{\link{dsr}},
\code{\link{sr_variance}}.

Other sr: 
\code{\link{as.sr}()},
\code{\link{confint.sr}()},
\code{\link{dsr}()},
\code{\link{is.sr}()},
\code{\link{plambdap}()},
\code{\link{power.sr_test}()},
\code{\link{predint}()},
\code{\link{print.sr}()},
\code{\link{reannualize}()},
\code{\link{sr_equality_test}()},
\code{\link{sr_test}()},
\code{\link{sr_unpaired_test}()},
\code{\link{sr_vcov}()},
\code{\link{sr}},
\code{\link{summary.sr}}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\concept{sr}
\keyword{htest}
