% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_clr.R
\name{start_dotnet_core_clr}
\alias{start_dotnet_core_clr}
\title{Start dotnet core runtime from an application base directory.}
\usage{
start_dotnet_core_clr(
  app_base_dir = NULL,
  runtime = "dotnet",
  version = "latest"
)
}
\arguments{
\item{app_base_dir}{defines the application base directory where the dlls are stored}

\item{runtime}{Select the shared runtime. The possible values are:
* \code{dotnet} - the \code{Microsoft.NETCore.App} shared runtime (default).
* \code{aspnetcore} - the \code{Microsoft.AspNetCore.App} shared runtime.}

\item{version}{Select the shared runtime version. \code{latest} by default. For more details see details section}
}
\description{
Start dotnet core runtime from an application base directory.
}
\details{
Starts a dotnet core runtime from an application base directory.
If the application base directory isn't self contained, the runtime and the version can be specifed.
For the version we apply the following heuristic:
If the value 'latest' or NULL is defined we gets the latest version installed for the given runtime.
Otherwise we try to match the version in this order:
\itemize{
\item Perfect match
\item Higher Major.Minor version found
\item Higher Major version found. This laste try cannot gurantee some breaking changes or other border effects.
}
}
