% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.r
\name{is.sr}
\alias{is.sr}
\title{Is this in the "sr" class?}
\usage{
is.sr(x)
}
\arguments{
\item{x}{an object of some kind.}
}
\value{
a boolean.
}
\description{
Checks if an object is in the class \code{'sr'}
}
\details{
To satisfy the minimum requirements of an S3 class.
}
\examples{
rvs <- as.sr(rnorm(253*8),ope=253)
is.sr(rvs)
}
\seealso{
sr

Other sr: 
\code{\link{as.sr}()},
\code{\link{confint.sr}()},
\code{\link{dsr}()},
\code{\link{plambdap}()},
\code{\link{power.sr_test}()},
\code{\link{predint}()},
\code{\link{print.sr}()},
\code{\link{reannualize}()},
\code{\link{se}()},
\code{\link{sr}},
\code{\link{sr_equality_test}()},
\code{\link{sr_test}()},
\code{\link{sr_unpaired_test}()},
\code{\link{sr_vcov}()},
\code{\link{summary.sr}}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\concept{sr}
