% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharpeR.r
\name{SharpeR-NEWS}
\alias{SharpeR-NEWS}
\title{News for package 'SharpeR':}
\description{
News for package 'SharpeR'

\newcommand{\CRANpkg}{\href{http://CRAN.R-project.org/package=#1}{\pkg{#1}}}
\newcommand{\SharpeR}{\CRANpkg{SharpeR}}
}
\section{Changes in \SharpeR{} Version 1.1.0 (2016-03-14) }{

\itemize{
\item fix sr_vcov on array input.
\item add SRIC method.
\item add SRIC to print.sropt.
\item change predint output to matrix.
}
}

\section{Changes in \SharpeR{} Version 1.0.0 (2015-06-18) }{


\itemize{
\item sane version numbers.
\item unpaired k sample test of Sharpe.
\item rely on same for unpaired 2 sample test.
\item prediction intervals for Sharpe based on upsilon.
\item more tests.
}
}

\section{Changes in \SharpeR{} Version 0.1501 (2014-12-06) }{


\itemize{
\item fix inference of mark frequency from e.g. xts objects.
\item add rlambdap.
}
}

\section{Changes in \SharpeR{} Version 0.1401 (2014-01-05) }{


\itemize{
\item fix second moment asymptotic covariance.
\item add confidence distribution functions for sr, sr.opt.
}
}

\section{Changes in \SharpeR{} Version 0.1310 (2013-10-30) }{


\itemize{
\item inverse second moment asymptotic covariance.
}
}

\section{Changes in \SharpeR{} Version 0.1309 (2013-09-20) }{


\itemize{
\item spanning/hedging tests.
\item sr equality test via callback variance covariance computation.
\item split vignette in two.
}
}

\section{Changes in \SharpeR{} Version 0.1307 (2013-05-30) }{


\itemize{
\item proper d.f. in sr objects with different nan fill.
\item restore vignette.
}
}

\section{\SharpeR{} Initial Version 0.1306 (2013-05-21) }{

\itemize{
\item put on CRAN
}
}

