\name{SharpeR-NEWS}
\alias{SharpeR-NEWS}
\title{News for package 'SharpeR':

\newcommand{\CRANpkg}{\href{http://CRAN.R-project.org/package=#1}{\pkg{#1}}}
\newcommand{\SharpeR}{\CRANpkg{SharpeR}}}
\description{
  News for package 'SharpeR':

  \newcommand{\CRANpkg}{\href{http://CRAN.R-project.org/package=#1}{\pkg{#1}}}
  \newcommand{\SharpeR}{\CRANpkg{SharpeR}}
}
\section{Changes in \SharpeR{} Version 0.1309 (2013-09-20) }{
  \itemize{ \item spanning/hedging tests. \item sr equality
  test via callback variance covariance computation. \item
  split vignette in two. }
}

\section{Changes in \SharpeR{} Version 0.1307 (2013-05-30) }{
  \itemize{ \item proper d.f. in sr objects with different
  nan fill. \item restore vignette. }
}

\section{\SharpeR{} Initial Version 0.1306 (2013-05-21) }{
  \itemize{ \item put on CRAN }
}

